+++ src/backend_new/core/workflow.py
@@ -13,9 +13,8 @@ from backend_new.extractors.geniusextractor import GeniusExtractor
 from backend_new.utils.functions.other import contains_japanese
 from backend_new.utils.functions.filesystem import read_json_file, write_json_file, read_config, load_env_file
 
-from backend_new.utils.exceptions import DataMismatchError
-from backend_new.utils.structures import SongContext
-from backend_new.utils.constants import TEMP_DIR
+from backend_new.utils.classes.exceptions import DataMismatchError
+from backend_new.utils.paths import TEMP_DIR
 from backend_new.utils.classes.dataclasses import SongContext, UIPromptRequest
 from backend_new.utils.classes.exceptions import DataMismatchError
 from backend_new.utils.paths import TEMP_DIR
